// msquare/fusion admin/fusionadmin/src/app/core/models/auth.models.ts
export class User {
  id: number = 0;
  password: string = '';
  photo: string = '';
  email: string = '';
  role: String = '';
  roleProfile: String = '';
  accessToken: string = '';
  refreshToken: string = '';
  name: string = '';
  companyId: string = '';
  issuperadmin: boolean = false;
  rememberMe: boolean = false;
  mobile: number = 0;
  permission:any = [];
  designation:any = {};
}